ignore error from mid function and fix admin link in report
At times this is passing a negative value as the third argument to the
mid function. I did not look at what is going on exactly, but for now
added a try/catch block to move past links causing this issue.

The links created to take the user to the site manager were for a much
older version of Mura and have been updated to work correctly.